refactor: fix cp, brotli is file exist check, add isExist function
