+++ src/utils/index.js
@@ -4,3 +4,4 @@ export { getByeMsg } from './getByeMsg.js';
 export { parseArgs } from './parseArgs.js';
 export { getValueByArg } from './getValueByArg.js';
 export { coloredLog } from './getColoredLog.js';
+export { isExist } from './isExist.js';
